Corrected models, applied migrations
